Introduce builder pattern for metrics
This commit introduces a builder pattern for metric construction.
The primary ambition here is to support adding and removing
metric tags hosted inside the metric in-flight.

I'm not _super_ stoked with the way Rust's builder looks acts in
terms of type-safety but it's better--I think--than fiddling around
inside of a type.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>