Adjust away from Builder
The builder notion didn't sit right with me, especially since bucket
was going to have to fiddle inside of Metric anyway (for the time
being). To that end, the builder's good things have been moved into
convenience methods on the Metric and the Metric::new is adjusted
to be a little less dorky.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>