Metrics now hold their own metadata
This commit moves the tags into Metric and, for the most part,
gets us out of fiddling with tags in the sinks. Wavefront is
a complication--because it emits global metrics which should be
the responsibility of a background producer or something and
because of a complication with timers / histograms. The latter
is captured in #127

Signed-off-by: Brian L. Troutwine <blt@postmates.com>