Metrics have metadata (#129)
* Introduce builder pattern for metrics

This commit introduces a builder pattern for metric construction.
The primary ambition here is to support adding and removing
metric tags hosted inside the metric in-flight.

I'm not _super_ stoked with the way Rust's builder looks acts in
terms of type-safety but it's better--I think--than fiddling around
inside of a type.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add documentation for new MetricsBuilder

This commit adds docs for MetricsBuilder--on account of how its
use may not be super clear--and these docs have tests. Fun!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Adjust away from Builder

The builder notion didn't sit right with me, especially since bucket
was going to have to fiddle inside of Metric anyway (for the time
being). To that end, the builder's good things have been moved into
convenience methods on the Metric and the Metric::new is adjusted
to be a little less dorky.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Metrics now hold their own metadata

This commit moves the tags into Metric and, for the most part,
gets us out of fiddling with tags in the sinks. Wavefront is
a complication--because it emits global metrics which should be
the responsibility of a background producer or something and
because of a complication with timers / histograms. The latter
is captured in #127

Signed-off-by: Brian L. Troutwine <blt@postmates.com>