Bump cernan to 0.4.2
This release incorporates several changes in many areas. In particular:

  * [EXPERIMENTAL] cernan is now able to forward on to other cernan instances
  * [STABLE] all metrics carry their timestamps, including histograms / timers
  * [STABLE] QOS sampling is now done by random choice in the flush interval
  * [EXPERIMENTAL] more file source changes have landed, improving use

The main show here is the sampling. In particular, the QOS sampling now
is done such that, if QOS is Q and the flush interval is F then there will
be no more than ceil(F/Q) points in the interval F. If there are less than
ceil(F/Q) points in F then all points will be shipped. If there are more,
the sampling will be done on a random basis to avoid introducing systematic
artifacts into the time series.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>