federation_receiver modifies metric metadata (#130)
* federation_receiver modifies metric metadata

As of this commit the federation_receiver merges its local tags onto
the top of the incoming metrics. This means that transmitters can
ship metrics with tags knowing they'll be preserved but the
receiver can mark a metric as having come through itself.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Modify equality for Metrics

We need Metrics to be considered distinct in the Bucket if the
tags of the Metric are distinct. To do that we define PartialOrd
for the Metric to take into account its tags. This is a little
long-winded and I'm open to faster implementations. I suppose
if we could figure out some order-independent code could keep
that when we modified tags.

Dunno!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>