Introduce an algebra of metrics
In order to support all metrics, including timers and histograms
carrying around their tags we have to be able to manipulate metrics
as if they were subject to some kind of algebra. In particular, we need
+ and =.

The nifty trick that's been done here--which, long term, means we're
capable of more sophisticated analysis of Metrics--is that each MetricKind
is now backed by a CKMS. The benchmark needle is lightly changed but because
of the CKMS doing ellision and its lack of need do a full binary search
we end up basically where we were.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>