Algebra of metrics (#131)
* Introduce an algebra of metrics

In order to support all metrics, including timers and histograms
carrying around their tags we have to be able to manipulate metrics
as if they were subject to some kind of algebra. In particular, we need
+ and =.

The nifty trick that's been done here--which, long term, means we're
capable of more sophisticated analysis of Metrics--is that each MetricKind
is now backed by a CKMS. The benchmark needle is lightly changed but because
of the CKMS doing ellision and its lack of need do a full binary search
we end up basically where we were.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Integrate clippy and fmt changes

This commit includes no correctness modifications but it does
get cernan back on the even keel of being fmt and clippy compliant.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Rely on metric tags in wavefront, always

This commit corrects the timer and histogram aggregation reports
to use the tags that come out of the metrics with that kind.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Switch quantiles dep to 0.2.0

Previously this branch was pegged at a working version of the quantiles
library. Now that that project has released 0.2.0 we're off tracking
the branch.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>