Implement bin width scale change
This commit makes it possible to configure a bucket with a bin-width
greater than one, which was the previous hard-coded assumption. The
Metric type has grown a 'within' function that allows you to say
whether two metrics are within some interval, with the time of the
metric called assumed to be in the middle of the interval.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>