In-progress log additions
This commit includes more logging in vital places and begins to
restructure the codebase to make this more convenient. It will
_not_ compile as I was flying without a compiler when I wrote
this and had to get off the plane before finishing in a sensible
state. Still, it's a start.

The ambition here is to make it easier to track down quirks with
cernan when run for long periods under high load. We've observed
'freezes' in reporting that are not easily explained.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>