Last work to get compiling cernan
This commit includes the last of the fiddling needed to get cernan
back into a state where it will compile. server.rs is now totally
changed into the source module, an improvement to the codebase's
readability. I am concern at how much goofy cloning the sources
do of their tags but that's an ongoing concern which is now made
more obvious.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>