Force panics to be aborts
This is not a long-term solution but it does mean that sub-thread
panics will not be allowed to keep cernan running. We need to have
some way of gracefully shutting down but since the mpsc is the
only place of concern _and_ there's no restarting yet, we're sorta
in the clear.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>