Scalable bins (#134)
* Remove QOS concept

As an experiment the QOS concept for wavefront has failed. We're
going, instead, to make the bin size--currently pegged at 1 second--
adjustable for each sink.

First step in that work is toasting QOS.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Implement bin width scale change

This commit makes it possible to configure a bucket with a bin-width
greater than one, which was the previous hard-coded assumption. The
Metric type has grown a 'within' function that allows you to say
whether two metrics are within some interval, with the time of the
metric called assumed to be in the middle of the interval.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make bin_width configurable by sink

As of this commit it's possible to configure the bin width for
both console and wavefront sinks independently of one another.
There remains README updates to document the changes here.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update README to reflect new reality

The nice thing about ditching the QOS concept is that it's much
simpler to explain. Benchmarking shows we got dinged on performance
somewhat so we'll need to address that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add tests to demonstrate counterexample

This commit adds a shuffling test to more easily demonstrate the
overlapping bin failure shown very seldomly by the other sampling
test.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct overlapping interval defect

This commit adjusts the 'within' determination to avoid the previously
identified overlapping interval issue. Happily, the fix is both simpler
to understand and a touch faster.

We now divide every timestamp by the bin width and take the floor. Boom.

This commit also includes code to read QC test adjustments from
environment variables. I've submitted a patch upstream to make this
convenient for all our QC tests: https://github.com/BurntSushi/quickcheck/pull/148

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* hist -> h