More Logging (#136)
* In-progress log additions

This commit includes more logging in vital places and begins to
restructure the codebase to make this more convenient. It will
_not_ compile as I was flying without a compiler when I wrote
this and had to get off the plane before finishing in a sensible
state. Still, it's a start.

The ambition here is to make it easier to track down quirks with
cernan when run for long periods under high load. We've observed
'freezes' in reporting that are not easily explained.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Further progress on compilation

This commit does not get cernan to the point where it will build
but it does move us closer to that goal. There's a fair bit of
small tinker work that needs doing.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Last work to get compiling cernan

This commit includes the last of the fiddling needed to get cernan
back into a state where it will compile. server.rs is now totally
changed into the source module, an improvement to the codebase's
readability. I am concern at how much goofy cloning the sources
do of their tags but that's an ongoing concern which is now made
more obvious.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Beginning of the integration of slog

This commit starts the integration of slog, down in terms of user
control. There's a crash at startup localized in graphite / statsd
sources that I'm unsure of presently. Already bound port issue.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Revert to fern, fix elapsed time logging

slog looks cool but it's documentation is scarce and it appears
that you have to thread a 'log' reference through your codebase.
This is orthogonal to our immediate need, though slog _does_
offer capability that would eventually be useful.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Force panics to be aborts

This is not a long-term solution but it does mean that sub-thread
panics will not be allowed to keep cernan running. We need to have
some way of gracefully shutting down but since the mpsc is the
only place of concern _and_ there's no restarting yet, we're sorta
in the clear.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add line number info

This commit adds line numbers into the log lines. There's some question
about the format being good or not, but Tom gets to do that work and so
I just get to toss it over the fence. Feels nice.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>
Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

* Alter logging message in receiver.rs

"[receiver | {}]" -> "[{}]"

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Re-introduce source startup logs

At some point the logs indicating that statsd/graphite sources
are starting were removed. This commit re-introduces those.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>