Fast-forward receiver to the end of index files (#137)
This commit fast-forwards the receiver to start reading at the
end of its index files. This avoids a case where mpsc restarts
and writes_to_read is 0 when the on-disk number of writes to
read is arbitrarily larger than 0.

Ideally we'd be able to restart the receiver where we left off.
This is a temporary solution. Achieving that will take a bit of
rework.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>