Break up config::Args into sub-structures
This commit starts on down the road of having a Filter concept. To
make that happen effectively we're going to need to have more
specific notions of configuring individual sources / sinks. To that
end, sources / sinks now are obligated to provide some kind of
Config object so that config.rs can stuff values down. A handy
side-effect is that we no longer have to copy config::Args.

This is a situation where inheritance would come in handy. Oh well.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>
Signed-off-by: Brian L. Troutwine <blt@postmates.com>