Roll our own file-watchers
It does not seem that notify is tested at high load, nor, I guess
is inotify used in this fashion really appropriate. What we care
about is reading large blocks of text off-disk with minimal latency,
not necessarily knowing when every little thing happens.

The good news is the codebase is now more direct and has the same
behaviour cross-platform. All we require is that the filesystem
support the 'created' metadata for files.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>