Backpressure and better file watch (#146)
* Cut of explicit backpressure notion between in the sinks

Previously we'd go as fast as possible in sinks pulling values off
the channels. This was a silly goof. It's entirely likely that the
far side will not be ready to accept what we've buffered and we'll
grow the buffer. This commit caps that buffer growth.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Roll our own file-watchers

It does not seem that notify is tested at high load, nor, I guess
is inotify used in this fashion really appropriate. What we care
about is reading large blocks of text off-disk with minimal latency,
not necessarily knowing when every little thing happens.

The good news is the codebase is now more direct and has the same
behaviour cross-platform. All we require is that the filesystem
support the 'created' metadata for files.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Properly handle errors

This commit ensures that we correctly handle absent files and other
such unfortunate circumstances.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Use std_dev, std_ino to id a file

This ought to be the most portable method for determining if the
file that we have open is in fact the file that we want to have
open. All other methods used so far are not portable enough.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make the firehose configurable

This commit makes the knobs present in the firehose sink available
through configuration.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update documentation

Signed-off-by: Brian L. Troutwine <blt@postmates.com>