Bump cernan to 0.4.7
This release contains two key features. The first, and most immediately
important, are adjustments to the file watching subsystem. We no longer
use filesystem events and prefer to do periodic polling. This resolves
overload of the OS notify mechanism on systems with very fast logs. As
well, this release includes the start of the scripting filter mechanism
though it is disabled.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>