Introduce Explicit Forwards, Filters (#149)
* Initial integration of lua extension

This commit introduces the use of lua in the filter concept,
allowing users to program modifications of in-flight metrics.
As of this commit nothing _super_ useful is going on but we
have some progress.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce programmable filters

This commit removes the special-purpose filters from the last commit
and makes it possible to program them directly with lua. There is no
good configuration to back this up. That comes next.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce more explicit form of configuration

This commit introduces a more explicit form of cernan configuration
which allows for declaring 'forwards' between sources, filters and
sinks. The big change here is to the testing code in config.rs.
The next commit will update cernan.rs to use this information.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Feed new configuration into into cernan binary

This commit feeds the new, more explicit configuration into the
cernan binary and down through the tests. What's especially fancy is
you can now set up a programmable filter and break the all sources
to all sinks, which resolves #39.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Further work on the programmable filter

Now configuration detail has been threaded through the system and
some work on properly catching errors has been done. There's a little
bit of 'leakage' going on, which needs investigation. An example:

Flushing metrics: 2016-11-29T22:41:32.998090+00:00
  counters:
    cernan.graphite.packet: 198
    cernan.graphite.packet: 1542
    cernan.graphite.packet: 1582
    cernan.graphite.packet: 1586
    cernan.graphite.packet: 1604
    cernan.graphite.packet: 1334
  gauges:
  raws:
    collectd.const: 337964
    collectd.const: 338730
    collectd.const: 339517
    collectd.const: 340305
    collectd.const: 341100
    collectd.const: 341880
    llrv.graphite.const: 337964
    llrv.graphite.const: 338730
    llrv.graphite.const: 339517
    llrv.graphite.const: 340305
    llrv.graphite.const: 341100
    llrv.graphite.const: 341880
  histograms:
  timers:
Flushing metrics: 2016-11-29T22:41:32.998676+00:00
  counters:
    cernan.graphite.packet: 2
  gauges:
  raws:
    collectd.const: 341881
    llrv.graphite.const: 341881
  histograms:
  timers:

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Modify firehose output to look more like heka

This commit modifies the fields we output to firehose to have more
of the heka flavor. We also add more in the way of tags to the payloads
as well as a UUID.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Allow tests to pass in Jenkins

This comments out some tests in programmable_filter and does not
replace them with integration tests. That'll come.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Open up multiple statsd / graphite sources

As of this commit the only source that does not allow multiples is
the federation receiver. Internal use at Postmates has suggested it'd
be very helpful to have multiple same-protocol sources running on
different ports for alternate routing and, eventually, filtering.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Mark work as being part of 0.4.8-pre

In the past we always kept on with the same version of cernan as had
just been landed into stable. This could cause confusion when doing
prototype work. This is a little nicer, I think, and may become
part of the standard flow.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update documentation

This commit includes documentation updates with regard to the new
explicit forwarding feature.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update README per John's feedback

Signed-off-by: Brian L. Troutwine <blt@postmates.com>