Add self-telemetry to wavefront sink
This commit introduces telemetry to the wavefront sink on the
number of delivery attempts that have been made as well as on the
"time spread" of metrics. The "time spread" gives us a histogram
of the difference between time::now() and all the metrics which
are being shipped. This _should_ range from 0 to flush-interval.

As delivery attempts grows that gives us a sense of being backed-up
and time spread lets us know what the wall-clock delay is.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>