Adjust how gauge times are maintained across resets (#151)
Here's a fun one! By retaining gauges across resets we've effectively
said that the time value of that gauge is as useful as its value. This
is not true. The time value _ought_ to be reset even while the value
persists.

Not doing so throws off the wavefront sink self-telemetry.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>