Bump to cernan 0.4.10
This release fixes a long-standing and never before recognized
issue with the way gauges' times were reset on bucket reset.
Previously, the last gauge's time was retained across a reset until
a new point came in. This works for relatively well-trafficed series
but does not hold up for sparse series of DeltaGauges.

We now reset times on bucket resets.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>