Bubble up errors from filters gracefully
This commit introduce a Result return from the filter as well as
an error type. One bad habit cernan has is that many of its APIs
assume success. This is a result of my being fairly inexperienced
with Rust when I started cernan. Filter now makes its own error
type and returns Result instead of failing silently.

Special thanks to John K. for pointing this out as a defect.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>