Stabilize Programmable Filters (#153)
* Move timers to top-level sinks, adapt lua API

This commit is a WIP but what it does is ensure that the executable
only sends flushes to source -> FOO desintations. This is needed
because we've always assumed that flush pulses would occur inline
with metric streams.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Integration test the programmable filter, expand API

This commit greatly expands the API of the programmable filter,
allowing the user to manipulate tag maps and add new metrics or
tags to the payload. A host of integration tests demonstrate
the good working order of this.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Extract a whole bunch of unwraps

Previously we used unwrap to crash cernan if a user didn't provide
enough arguments to a function. I bet there's a nicer lua way to
guard against this, but I surely couldn't find it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Emit flushtimers only to sinks

There's a fair bit more research that needs to be done to decide
exactly how to filter flushes through an arbitrary graph. Until
that research is done we'll keep the behaviour we've had all
this time.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update the README for filters

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* More comprehensive collectd scrub, test

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Bubble up errors from filters gracefully

This commit introduce a Result return from the filter as well as
an error type. One bad habit cernan has is that many of its APIs
assume success. This is a result of my being fairly inexperienced
with Rust when I started cernan. Filter now makes its own error
type and returns Result instead of failing silently.

Special thanks to John K. for pointing this out as a defect.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Expand FilterError to require a metric

This addition to the FilterError makes it possible for us to
self-telemeter on failure of a filter. This is a hard failure
detector. We can extend the contract with the lua programs to
telemeter on other kinds of failure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>