Remake the hard distinction between gauges / delta-gauges (#155)
We previously blurred the line between gauges and delta-gagues--or
increment only counters--and that turned out not to be great in a
subtle way. Say you have a gauge-only series that reports every
so often. What does cernan do? Well! It turns out it will persist
the last value. Worse, a delta-gague series will reset every bin
interval to sum up when it should persist across intervals.

Womp womp

Signed-off-by: Brian L. Troutwine <blt@postmates.com>