Callgrind Performance Improvements (#156)
* Begin work on callgrind guided optimizations

This series of commits is intended to bump cernan's performance by
eating up the low-hanging-fruit available to us. Callgrind's our
main jam right now.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce array-form of TagMap

In an attempt to drop the runtime penalty incurred by each metric having
its own hashmap of tags I've introduced a TagMap based on a vec of tuples.
I'm not sure if it's a good idea yet, but we'll find out!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct doctests

The move to a custom TagMap has replaced some of the Into conversions
that HashMap had naturally. Womp womp

This commit repairs the doctests that required those conversions.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove shift goofiness in mpsc::sender

For a good long while we've been shifting the payload length of
the mpsc sender onto the front of every payload. Obviously that's
not the best thing in the world since you have to shift the payload
not just once but four times.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>