Bump cernan to 0.4.11
This release stabilizes the programmable filters, making it possible
to manipulate telemetry and logs in-flight with user-written lua
code. This is a big accomplishment, one of the long-term goals of
cernan is now achieved.

This release also fixes a bug in the interpretation of gauges by
making a hard distinction between delta-gauges and gauges. Previously
their individual behaviours were somewhat confused.

Lastly, this release introduces some performance improvements driven
by callgrind profiling of cernan. There's more that can be done--
especially around metrics which are mostly the same internally--but
this was all low-hanging fruit.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>