Correct obligatory 1ms wait in time::delay
Well, guess what? It turns out that above a certain, very small,
PPS that cernan would back up. Not irrevocably but noticably and the
more the PPS got cranked the worse things would become.

Why's that? There was a change made to time::delay a little bit ago
that, through sheer ignorance, removed a special consideration for
an input of 0. Without this special consideration time::delay will
delay for 1ms. With the previous special consideration it delays
for 0ms.

Turns out, that will cause a whole lot of backup.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>