Fix Processing Delay (#159)
* Avoid allocations, multiply error in time::delay

The programmable_filter used to do a fair bit of extra allocation.
This has been corrected by adjusting who owns what and judicious
use of Box. Further, it was possible for time::delay to overflow
its 32 bits if attempts were too high. Max wait is now 512ms and
max attempts is now 9.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add more debugging info to wavefront sink

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Be sure to emit delta_gauges

Due to an oversight delta_gauges are not flushed from the WF sink

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Split the notion of gauge / delta-gauge time spreads

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Emit counts for WF time_spreads

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Fiddle with Jenkinsfile to speed up exploration

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove excess debugging

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct obligatory 1ms wait in time::delay

Well, guess what? It turns out that above a certain, very small,
PPS that cernan would back up. Not irrevocably but noticably and the
more the PPS got cranked the worse things would become.

Why's that? There was a change made to time::delay a little bit ago
that, through sheer ignorance, removed a special consideration for
an input of 0. Without this special consideration time::delay will
delay for 1ms. With the previous special consideration it delays
for 0ms.

Turns out, that will cause a whole lot of backup.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Repair Jenkinsfile

In our rush to experiment we needed faster builds. However, we do
want cernan to be actively tested under the beta channel so we
know what's coming down the pipe.

This commit replaces the beta tests.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove excess debugging

Issuing a debug every time we get a new point into wavefront is
not a great idea. There's a CPU cost to it, especially since we're
using the debug print to do so.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>