Avoid further allocations in sending over multiple chans
In a common case we have only one chan to pass data onto. There's
no need for us to do an extra allocation. Even when there _is_ more
than one chan in a pipeline we can avoid allocating +1 copy of the
data by giving ownership of the original to the last chan in the
vec.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>