Introduce native protocol, CPU resource improvements (#167)
This commit includes a number of broad changes to cernan with the
intention of drastically reducing the CPU load that it requires
at high levels of ingestion. This has been done through smarter
use of memory--we've reduced small allocations through the
judicious application of Arc--and through a much more efficient
protobuf based ingestion protocol. Text based sinks have been
improved as well through push-based building of their strings.

The self-telemetry of the wavefront sink is reduced. The problems
that needed addressing have been and the CPU drain was non-trivial.
A cheap method for self-telemetry should be investigated.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>