Introduce an InfluxDB sink (#158)
This commit introduces an InfluxDB sink. It aggregates and ships
those aggregates on to InfluxDB using the UDP / line protocol.
Aggregations are done according to the method given in the project
README. The sink could be rewritten to simply store and forward,
which might be interesting. I'm open to that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>