Begin to work the native source
There's some goofy stuff going on in the protobuf implementation
that needs correcting, or something. The statsd source--a text
protocol--barely moves the needle but this here protobuf based
approach eats up 10%. This is a result of many small allocs that are
endemic to protobuf implementations. A slab allocator for cernan would
not be a bad idea. The majority of the allocations are short-lived
containers.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>