Allocate Less (#168)
* Reduce the allocations of statsd

There's nifty tricks we can do to avoid .lines() String allocations
if there can be a BufReader applied to the source of a thing.
Unfortunately, UdpSocket doesn't implement Read so without being
more clever than I am presently this is what I've got for statsd.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Reduce graphite allocations

This commit improves the graphite subystem by passing a pre-allocated
string and metric into parse_graphite. We are no longer allocating
a string per line as before and do not have to pay the penalty for
overlaying tags all the time.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Begin to work the native source

There's some goofy stuff going on in the protobuf implementation
that needs correcting, or something. The statsd source--a text
protocol--barely moves the needle but this here protobuf based
approach eats up 10%. This is a result of many small allocs that are
endemic to protobuf implementations. A slab allocator for cernan would
not be a bad idea. The majority of the allocations are short-lived
containers.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>