Extend the native protocol to have explicit persist option
In #172 we identified that cernan's method of persisting statsd gauges
is a little weak. The ambition for the native protocol is to reflect
all of the subtlties of the statsd protocol--which is by far the most
expansive of cernan's sources--while making them explicit and clear.
The intention--noted in #171--is that this should drag cernan's internal
model along with it.

The native protocol now has an explicit 'persisted' flag in Telemetry.
A telemetry object which has been tagged as persisted will persist
across binning windows. One which is not persisted will not.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>