Adjust cernan's internal model to reflect native protocol (#175)
* Adjust cernan's internal model to reflect native protocol

This commit moves cernan's internal model away from the explicitly
statsd inspired language toward cernan's native protocol. The native
protocol explicitly encodes the information around persistence and
aggregation that statsd does without the overloading of various
types. Future expansion is also now possible by extending the scope
of AggregationMethod.

Many areas have been touched but the broad shape of things has been
made more clear.

This resolves #172 and #171

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add fitness for purpose test, statsd gauge

This commit in particular focuses on demonstrating resolution of #172.
We hadn't _quite_ made it, as per the change in Telemetry. In particular
the inheritance of persist must not be an OR but should instead be a
set from the RHS.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>