Modify the way statsd gauges get parsed
We've moved to be more compatible with etsy/statsd by making all
gauges persistent by default. Any signed gauge value will be taken
as a SUM, any unsigned as a SET. It is possible to adjust the
persistence of gauge lines by modifying a delete-gauge variable
on the source configuration.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>