Resolve graphite growth problem, introduce hopper 0.1.2 (#181)
We noticed that there was something fishy going on with graphite.
In particular, the number of lines push to the programmable_filter
grew _way_ too quickly for the inputs. It turns out, that fancy
line buffer we pass in to avoid allocing small strings repeatedly
was not being cleared and so everything that came in would be
appended there.

This commit also includes the work done as a part of
postmates/hopper:#5 by updating to 0.1.2. All messages will be
delivered in order--as they should have been--and hopper is now
quite a bit faster to boot as it can keep cramming in low-memory
when there's space.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>