Bump to 0.5.0
This release is a long time coming and closes the venerable 0.4.x
series of development. Cernan has been put on a good foot for
being open-sourced, has received significant optimizations and
been given three new sinks and one new source. The changes are
as follows:

  * ccced687 :: Telemetry now avoids allocing a CKMS if not needed
  * 3d88add3 :: mpsc is split out into the open-source hopper
  * 3937bc45 :: native protocol is introduced, Arc sprinkled throughout
  * 22668f42 :: Experimental InfluxDB sink is introduced
  * 47467664 :: Numerous small allocations in text sources are removed
  * f48403a7 :: Experimental Prometheus sink is introduced
  * 4b609f33 :: Cernan's internal data-model becomes that of native protocol
  * b1d482ff :: Aggregation optimizations, wrt cache access

There were a few things I skipped in the above either because they
were later supplanted or because it was hard to fit the summary in
a short line. But that's the gist!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>