Get travis to check the format of branches (#200)
* Get travis in the business of checking formatting

This commit will configure travis.ci to check the formatting of
branches. It's a finicky thing -- and hopefully it won't scare
away contributors -- but it's good to have uniformity.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Pass automatic travis fmt check

The existing cernan codebase did not cleanly pass a cargo fmt check
on account of trailing whitespace and lines that were much too long.
This is corrected.

Travis builds are now _much_ longer than they were. This is unfortunate.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>