Update cernan to use serde 0.9 (#204)
* Introduce codecov.yml

We want to be sure that we keep code coverage of this project to
a certain level, 95%. Coverage is not correctness but it does
show whether we've probed sufficiently in test. We allow coverage
to drop by 1% per PR.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update cernan to use serde 0.9

This commit removes the special codegen step that we had to take with
rust < 1.15 and serde < 0.9. Now that 1.15 is Stable we can remove
this codegen step. Builds take about the same speed but the code
is a little nicer to interact with.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make a distinction between patch/project in codecov

The codecov configuration file now makes sure that patches are at
100% while the project as a whole may have coverage of 95%. This
meets current norms and demands very high coverage going forward.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>