Correct firehose sink  (#207)
* Avoid allocating a new lines list

This commit uses "drain" on the lines vec to avoid allocating a
new vec each time we go around to pull lines from files. This'll
cause no change in behaviour.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add a delay into the log pulling loop

We're spinning pretty quickly on systems that return OK when there's
no line in-buffer. That is, we're burning CPU spinning doing nothing.
This adds a time::delay into the loop, as per custom.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Report success when writing into delivery stream

To help debug an issue with delivery into Kinesis Firehose we're
adding a debug statement to the success branch. Hopefully this
will give us some hint of what's up.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct firehose loss of events

A regression was introduced in 0.5.0 that removed event retrieval
for the firehose sink. This is now corrected. We have re-introduced
the correct integration of the input logline into internal buffer
and have confirmed attempted transmission.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>